+++ client/src/api/content.js
@@ -1,0 +1,5 @@
+import { get } from './base'
+
+export async function getContent (contentId) {
+  return await get(`/content/${contentId}`)
+}