+++ src/Modules/Content/ContentGateway.php
@@ -2,13 +2,38 @@
 
 namespace Foodsharing\Modules\Content;
 
+use Carbon\Carbon;
+use DateTimeZone;
+use Foodsharing\Modules\Content\DTO\Content;
 use Foodsharing\Modules\Core\BaseGateway;
 
 class ContentGateway extends BaseGateway
 {
+    /**
+     * @deprecated use getContent instead
+     */
     public function get($id): array
     {
         return $this->db->fetchByCriteria('fs_content', ['title', 'body'], ['id' => $id]);
+    }
+
+    /**
+     * Returns the content with the specific id or null if the id does not exist.
+     */
+    public function getContent(int $id): ?Content
+    {
+        $content = $this->db->fetchByCriteria('fs_content', ['title', 'body', 'last_mod'], ['id' => $id]);
+
+        if ($content == null) {
+            return null;
+        }
+
+        $lastModified = $content['last_mod'] != null
+            ? Carbon::createFromFormat('Y-m-d H:i:s', $content['last_mod'], new DateTimeZone('Europe/Berlin'))
+                ->shiftTimezone(new DateTimeZone('UTC'))
+            : null;
+
+        return Content::create($content['title'], $content['body'], $lastModified);
     }
 
     public function getMultiple(array $ids): array